create the new node builder and add grid loader with node loader cuibtre
