+++ ai_gen/lib/features/node_view/presentation/cubits/node_view_cubit.dart
@@ -1,0 +1,24 @@
+import 'package:ai_gen/features/node_view/presentation/node_builder/node_builder.dart';
+import 'package:ai_gen/node_package/data/vs_subgroup.dart';
+import 'package:bloc/bloc.dart';
+import 'package:meta/meta.dart';
+
+part 'node_view_state.dart';
+
+class NodeLoaderCubit extends Cubit<NodeLoaderState> {
+  NodeLoaderCubit() : super(NodeViewInitial()) {
+    buildBlocks();
+  }
+
+  Future buildBlocks() async {
+    try {
+      emit(NodeViewLoading());
+
+      final List<VSSubgroup> nodeBuilder = await NodeBuilder().buildBlocks();
+
+      emit(NodeViewSuccess(nodeBuilder));
+    } catch (e) {
+      emit(NodeViewFailure("Failed to load blocks"));
+    }
+  }
+}
