--- ai_gen/lib/features/node_view/presentation/node_view_cubit/node_view_cubit.dart
@@ -1,8 +1,0 @@
-import 'package:bloc/bloc.dart';
-import 'package:meta/meta.dart';
-
-part 'node_view_state.dart';
-
-class NodeViewCubit extends Cubit<NodeViewState> {
-  NodeViewCubit() : super(NodeViewInitial());
-}
