--- ai_gen/lib/features/node_view/presentation/node_view_cubit/node_view_state.dart
@@ -1,6 +1,0 @@
-part of 'node_view_cubit.dart';
-
-@immutable
-sealed class NodeViewState {}
-
-final class NodeViewInitial extends NodeViewState {}
